Avoid unsigned ints in the IDL as kotlin treats them as experimental

+++ nimbus/src/evaluator.rs
@@ -130,7 +130,8 @@ pub fn evaluate_enrollment(
 ///
 /// An error could occur if something goes wrong while sampling the ratios
 fn choose_branch<'a>(slug: &str, branches: &'a [Branch], id: &str) -> Result<&'a Branch> {
-    let ratios = branches.iter().map(|b| b.ratio).collect::<Vec<_>>();
+    // convert from i32 to u32 to work around SDK-175.
+    let ratios = branches.iter().map(|b| b.ratio as u32).collect::<Vec<_>>();
     // Note: The "experiment-manager" here comes from
     // https://searchfox.org/mozilla-central/rev/1843375acbbca68127713e402be222350ac99301/toolkit/components/messaging-system/experiments/ExperimentManager.jsm#469
     // TODO: Change it to be something more related to the SDK if it is needed
